Separate library matching code into class
* also add two params to influence library matching
and dependency walking
