+++ docs/Usage.md
@@ -67,5 +67,33 @@ Options:
                           Can be set using RDK_OCI_REGISTRY_CREDS environment
                           variable for security
 
+  -y, --yes                       Automatic yes to prompt
+
+  -n, --nodepwalking              Dependency walking and library matching is
+                                  active by default. Use this flag to disable
+                                  it. When enabled, the dependencies of all
+                                  libs indicated in gfxLibs and
+                                  pluginDependencies config, will
+                                  automatically also be added to the bundle.
+                                  Host or image version of library is decided
+                                  by libmatchingmode parameter below. This
+                                  logic can only work if a _libs.json file is
+                                  present with libs and apiversions info.
+
+  -m, --libmatchingmode [normal|image|host]
+                                  normal: take most recent library i.e. with
+                                  most api tags like 'GLIBC_2.4'.
+
+                                  image: always take lib from image rootfs, if
+                                  available in there.
+
+                                  host: always take host lib and create mount
+                                  bind. Skips the library from image rootfs if
+                                  it was there.
+
+                                  Default mode is 'normal'. When apiversion
+                                  info not available the effect is the same as
+                                  mode 'host'
+
   --help                  Show this message and exit.
 ```
